Fix remaining bugprone warnings (14 files)
- Rename reserved identifiers (_FillFunctionDrop, _STRINGS_H_, _S_CPP_)
- Add default: break to switch statements in frmNewMod
- Fix implicit widening of multiplication results with static_cast
- Add NOLINT for intentional branch clones with explanatory comments
- Move increment out of complex conditions in CUcsFile
- Fix too-small loop variable type in CSgaFile
- Handle unused return value in CResourceLoader

+++ src/cdms/Construct.cpp
@@ -734,8 +734,8 @@ void ConstructFrame::DoLoadMod(wxString sPath, eLoadModGames eGame)
         MD5Final(sModConfigKey, &md5ModName);
         for (int i = 0; i < 16; ++i)
         {
-            sModConfigKeyHex[i * 2] = "0123456789ABCDEF"[sModConfigKey[i] >> 4];
-            sModConfigKeyHex[(i * 2) + 1] = "0123456789ABCDEF"[sModConfigKey[i] & 15];
+            sModConfigKeyHex[static_cast<size_t>(i) * 2] = "0123456789ABCDEF"[sModConfigKey[i] >> 4];
+            sModConfigKeyHex[(static_cast<size_t>(i) * 2) + 1] = "0123456789ABCDEF"[sModConfigKey[i] & 15];
         }
         TheConfig->SetPath(wxString(wxT("/")).Append(sModConfigKeyHex));
